Clean NA-heavy columns from a Data Frame or Tibble

Remove columns whose proportion of NAs is higher than determined threshold. Setting threshold of 0.25 asks R to remove all columns with 25% or more NA values. Can be useful when dealing with large data frames, where many columns are redundant.

limpiar_na_cols(df, threshold)

Arguments

df
The Data Frame or Tibble object

threshold
Threshold of non-NA entries a column must exceed to be retained.

Value

Data Frame or Tibble with NA-heavy columns purged

Examples

limpiar_examples

#> # A tibble: 10 × 5
#>    doc_id author_name       mention_content       mention_url   platform_intera…
#>     <int> <chr>             <chr>                 <chr>         <lgl>           
#>  1      1 don_quijote       "mi amigo sancho es … www.twitter.… NA              
#>  2      2 sancho_panza      "RT mi amigo sancho … www.twitter.… NA              
#>  3      3 edmond_dantes     "@don_quijote no dig… www.twitter.… NA              
#>  4      4 el_sordo          "nos han metido en u… www.fakebook… NA              
#>  5      5 commander_miranda "nos han metido en u… www.fakebook… NA              
#>  6      6 robert_jordan     "   Lo q no tenemos … www.youtube.… NA              
#>  7      7 anselmo           "a mi es muy grave q… www.twitter.… NA              
#>  8      8 maria             "ayyy nooo @robert_j… www.twitter.… NA              
#>  9      9 pablo             "todos se unen a nue… www.instagra… NA              
#> 10     10 pilar             "a mi me gustaría qu… www.instagra… NA              

limpiar_examples %>% limpiar_na_cols(0.1)

#> # A tibble: 10 × 4
#>    doc_id author_name       mention_content                    mention_url      
#>     <int> <chr>             <chr>                              <chr>            
#>  1      1 don_quijote       "mi amigo sancho es un wn de vdd … www.twitter.com/…
#>  2      2 sancho_panza      "RT mi amigo sancho es un wn de v… www.twitter.com/…
#>  3      3 edmond_dantes     "@don_quijote no digas eso, tu am… www.twitter.com/…
#>  4      4 el_sordo          "nos han metido en una muy difici… www.fakebook.com…
#>  5      5 commander_miranda "nos han metido en una muy difici… www.fakebook.com…
#>  6      6 robert_jordan     "   Lo q no tenemos es tiempo.   … www.youtube.no.r…
#>  7      7 anselmo           "a mi es muy grave quitarle la vi… www.twitter.com/…
#>  8      8 maria             "ayyy nooo @robert_jordan 😢 😢 …  www.twitter.com/…
#>  9      9 pablo             "todos se unen a nuestro grupo ha… www.instagram.co…
#> 10     10 pilar             "a mi me gustaría quedarme un rat… www.instagram.co…