Annotate Texts for Parts of Speech Analysis

Annotate Texts for Parts of Speech Analysis

limpiar_pos_annotate(
  data,
  text_var,
  id_var,
  pos_model,
  in_parallel = FALSE,
  parse_text = FALSE,
  ...
)

Arguments

data

The data.frame or tibble object containing any texts that the user wishes to conduct parts of speech analysis on.

text_var

Any texts or sentences the user wishes to perform the parts of speech annotations on.

id_var

The document identifier, this can be something like universal_message_id or doc_number. The default is universal_message_id.

pos_model

A UDPipe model imported using limpiar_pos_import_model, and must be of class 'udpipe_model'.

in_parallel

A logical argument allowing the user to initiate parallel processing to speed the annotate function up. If set to TRUE, the function will select the number of available cores minus one, processing more efficiently(faster), leaving one core to manage other computations. The default is FALSE.

parse_text

Whether to perform dependency parsing on tokens. The default is set to FALSE. If the user wishes to perform parsing on tokens, they may do so by calling TRUE.

...

To enable the user to supply any additional arguments.

Value

Returns a data frame with documents broken up into both a token and sentence level, in addition to the existing variables present in data supplied to the function. The returned object contains the parts of speech annotations in CONLL-U formatting, where each row is an annotation of a word. To find out more on the formatting methods, read here. The additional arguments with tagged POS information are as follows:

paragraph_id: The identifier indicating the paragraph the annotated token is derived from.

sentence_id: Similar to paragraph_id but at a sentence level.

sentence: The sentence the annotated token is derived from.

token_id: Token index, integer starting at 1 for each new sentence. May be a range for multiword tokens or a decimal number for empty nodes.

token: The token(or word) being annotated for parts of speech.

lemma: The lemmatized version of the annotated token.

pos_tag: The universal parts of speech tag of the token. Here for more information.

xpos: The treebank-specific parts of speech tag of the token.

feats: The morphological features of the token, used for dependency parsing visualisations,.

head_token_id: Indicating what is the token id of the head of the token, indicating to which other token in the sentence it is related.

dependency_tag: The information regarding dependency parsing of tokens, displaying the type of relation the token has with the head_token_id. Here for more information.

Examples

data <- dplyr::tibble(text = tolower(stringr::sentences[1:100]),
document = 1:100)
model <- LimpiaR::limpiar_pos_import_model(language = "english")
#> Downloading udpipe model from https://raw.githubusercontent.com/jwijffels/udpipe.models.ud.2.5/master/inst/udpipe-ud-2.5-191206/english-ewt-ud-2.5-191206.udpipe to /var/folders/nc/y2v839d56c93m7bmzchn29sh0000gn/T//RtmpqgXnnN/english-ewt-ud-2.5-191206.udpipe
#>  - This model has been trained on version 2.5 of data from https://universaldependencies.org
#>  - The model is distributed under the CC-BY-SA-NC license: https://creativecommons.org/licenses/by-nc-sa/4.0
#>  - Visit https://github.com/jwijffels/udpipe.models.ud.2.5 for model license details.
#>  - For a list of all models and their licenses (most models you can download with this package have either a CC-BY-SA or a CC-BY-SA-NC license) read the documentation at ?udpipe_download_model. For building your own models: visit the documentation by typing vignette('udpipe-train', package = 'udpipe')
#> Downloading finished, model stored at '/var/folders/nc/y2v839d56c93m7bmzchn29sh0000gn/T//RtmpqgXnnN/english-ewt-ud-2.5-191206.udpipe'
annotations <- limpiar_pos_annotate(data = data,
                                   text_var = text,
                                   id_var = document,
                                   pos_model = model,
                                   in_parallel = FALSE,
                                   parse_text = TRUE)
#> Performing dependency parsing on tokens...
#> Parts of speech tagging in process...